pyproject.toml
[tool.poetry]
name = "nansen"
version = "0.1.0"
description = ""
authors = ["Alen <alen.kimov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
tqdm = "^4.66.1"
loguru = "^0.7.0"
aiohttp = "^3.8.0"
lxml = "^4.9.3"
python3-aioanticaptcha = "^1.2"
python-anticaptcha = "^1.0.0"
pydantic = "^2.1.1"
pyuseragents = "^1.0.5"
better-web3 = "^2.0.0"
imapclient = "^2.3.1"
beautifulsoup4 = "^4.12.2"
questionary = "^2.0.0"
aioimaplib = "^1.0.1"
aiohttp-socks = "^0.8.0"
httpx = "^0.24.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
